webapplicationformedicalimagingai-master/requirements.txt
Django==4.1.7
pre-commit
torch==1.13.1
numpy==1.24.2
matplotlib==3.7.0
monai==1.1.0
nibabel==5.0.1
djangorestframework==3.14.0
pytest==6.2.5
tensorflow==2.11.0
pandas==1.5.3
